+++ x/pool/handler.go
@@ -15,8 +15,8 @@ func NewHandler(k *keeper.Keeper) sdk.Handler {
 		case *types.MsgCreatePool:
 			res, err := msgServer.CreatePool(sdk.WrapSDKContext(ctx), msg)
 			return sdk.WrapServiceResult(ctx, res, err)
-		case *types.MsgModifyPoolWeight:
-			res, err := msgServer.ModifyPoolWeight(sdk.WrapSDKContext(ctx), msg)
+		case *types.MsgSetPoolWeight:
+			res, err := msgServer.SetPoolWeight(sdk.WrapSDKContext(ctx), msg)
 			return sdk.WrapServiceResult(ctx, res, err)
 		default:
 			return nil, sdkerrors.Wrapf(sdkerrors.ErrUnknownRequest, "unrecognized %s message type: %T", types.ModuleName, msg)
