+++ x/pool/keeper/keeper.go
@@ -46,12 +46,3 @@ func (k Keeper) isValidOperator(ctx sdk.Context, operatorAddress string) bool {
 	}
 	return false
 }
-
-func (k Keeper) GetParams(ctx sdk.Context) (params types.Params) {
-	k.paramSpace.GetParamSet(ctx, &params)
-	return params
-}
-
-func (k Keeper) SetParams(ctx sdk.Context, params types.Params) {
-	k.paramSpace.SetParamSet(ctx, &params)
-}
